Entry

Description

This is the main entry point into our logic, with everything already nicely in the world of Haskell, i.e. no JS dependencies.

Documentation

incredibleLogic :: Context -> Proof -> Either String Analysis

incredibleNewRule :: Context -> Proof -> Either String (Maybe Rule)