Lint

Description

This module is a linter for the input and output of our programs. This means it checks properties that, if they fail, are a bug somewhere in the program.

It should not be possible to trigger the linter via the usual interface (but easily so via the debugging interface).

The existence of this module means that Haskell’s type system is not perfect...

The name "linter" comes from GHC, which also has that.

Documentation

type Lints = [String]

lint :: Context -> Proof -> Lints

lintsToEither :: Lints -> Either String ()