ProofGraph

Description

The proof is (natuarlly) a directed graph. For full detail, one can view blocks, out-ports, connections and in-ports as nodes of the graph. This module realizes this view, and provides algorithms based on that.

This module plays with type families to get a very strongly typed graph, and it ties the know to have the graph as such on the heap, i.e. no maps to go to the successor.

Documentation

data NodeType

Constructors

BlockNodeType

OutPortNodeType

InPortNodeType

ConnNodeType

data NodeTag a where

Constructors

BlockNodeTag :: NodeTag BlockNodeType

OutPortNodeTag :: NodeTag OutPortNodeType

InPortNodeTag :: NodeTag InPortNodeType

ConnNodeTag :: NodeTag ConnNodeType

type family Succ a

Equations

Succ BlockNodeType = OutPortNodeType

Succ OutPortNodeType = ConnNodeType

Succ ConnNodeType = InPortNodeType

Succ InPortNodeType = BlockNodeType

type family Pred a

Equations

Pred BlockNodeType = InPortNodeType

Pred OutPortNodeType = BlockNodeType

Pred ConnNodeType = OutPortNodeType

Pred InPortNodeType = ConnNodeType

type family NodeKey a

Equations

NodeKey BlockNodeType = Key Block

NodeKey OutPortNodeType = PortSpec

NodeKey InPortNodeType = PortSpec

NodeKey ConnNodeType = Key Connection

type NodePath = [NodeUniq]

type NodeUniq = Int

data Node a

Constructors

Node

Fields

nodeType :: NodeTag a

nodeKey :: NodeKey a

nodePred :: [Node (Pred a)]

nodeSucc :: [Node (Succ a)]

nodeUniq :: NodeUniq

data ANode

Constructors

forall a . ANode (Node a)

type NodeMap a = Map (NodeKey a) (Node a)

type UniqMap a = IntMap (NodeKey a)

data Graph

Constructors

Graph

Fields

blockNodes :: NodeMap BlockNodeType

inPortNodes :: NodeMap InPortNodeType

outPortNodes :: NodeMap OutPortNodeType

connectionNodes :: NodeMap ConnNodeType

uniquesToBlock :: UniqMap BlockNodeType

uniquesToInPort :: UniqMap InPortNodeType

uniquesToConn :: UniqMap ConnNodeType

localHypNodes :: UniqSet

conclusionNodes :: [Node BlockNodeType]

ppGraph :: Graph -> String

mkNodeMap :: Ord (NodeKey a) => Int -> [Int -> Node a] -> NodeMap a

mkNodeMap' :: Ord (NodeKey a) => Int -> [Int -> Node a] -> (NodeMap a, UniqMap a)

proof2Graph :: Context -> Proof -> Graph

calcCycle :: Node a -> UniqSet -> [NodePath]

Finds a list of cycles from the given node to itself, ignoring the nodes from stopAt along the way, and returns such paths.

calcNonScope :: Node a -> UniqSet -> UniqSet -> IntMap NodePath

Starting with the given node, this moves forward from the node, and then backwards, always ignoring the nodes listed in stopAt.

It returns a map from all visited nodes, with the path from the start node to the visited node as values.

The nodes mentiond in stopAt are _not_ included in the returned map.

type NonScopeM = State (IntMap NodePath)

type NonScopeDeferM = WriterT [NonScopeM ()] NonScopeM

calcSCC :: Node a -> UniqSet

backwardsSlice :: [Node a] -> [NodeUniq]

toBlockNodeKey :: Graph -> NodeUniq -> Maybe (Key Block)

toConnKey :: Graph -> NodeUniq -> Maybe (Key Connection)

toInPortKey :: Graph -> NodeUniq -> Maybe PortSpec

type UniqSet = IntSet

type MarkM = State UniqSet

execMarkM :: MarkM a -> UniqSet

evalMarkM :: MarkM a -> a

markAndFollow :: Monoid m => Int -> MarkM m -> MarkM m